+++ ahachul_backend/src/test/kotlin/backend/team/ahachul_backend/api/train/adapter/in/TrainControllerDocsTest.kt
@@ -79,7 +79,7 @@ class TrainControllerDocsTest : CommonDocsTestConfig() {
     fun getTrainCongestionTest() {
         // given
         val response = GetCongestionDto.Response(
-            trainNo = 2023,
+            trainNo = "2023",
             congestions =
                 listOf(
                     GetCongestionDto.Section(
@@ -97,8 +97,8 @@ class TrainControllerDocsTest : CommonDocsTestConfig() {
         // when
         val result = mockMvc.perform(
             get("/v1/trains/real-times/congestion")
-                .queryParam("subwayLineId", "1")
-                .queryParam("trainNo", "1123")
+                .queryParam("subwayLineId", "2")
+                .queryParam("trainNo", "2023")
                 .header("Authorization", "Bearer <Access Token>")
                 .accept(MediaType.APPLICATION_JSON)
         )
@@ -119,7 +119,7 @@ class TrainControllerDocsTest : CommonDocsTestConfig() {
                     ),
                     responseFields(
                         *commonResponseFields(),
-                        fieldWithPath("result.trainNo").type(JsonFieldType.NUMBER).description("열차 식별 번호"),
+                        fieldWithPath("result.trainNo").type(JsonFieldType.STRING).description("열차 식별 번호"),
                         fieldWithPath("result.congestions[]").type(JsonFieldType.ARRAY).description("열차 혼잡도 배열"),
                         fieldWithPath("result.congestions[].sectionNo").type(JsonFieldType.NUMBER).description("칸 번호"),
                         fieldWithPath("result.congestions[].congestionColor").type(JsonFieldType.STRING).description("칸 혼잡도 정도")
